requirements.txt
fastapi
uvicorn[standard]
redis
httpx
python-multipart
websockets
pillow
requests
pynput
scapy
